+++ src/command/tools/fmtOutPutReport.ts
@@ -1,4 +1,5 @@
 import * as vscode from 'vscode';
+import { getFormatConfig } from '../../configUI';
 import { EFormatChannel } from '../../globalEnum';
 import { FormatCore, FormatCoreWrap } from '../../provider/Format/FormatProvider';
 
@@ -12,6 +13,7 @@ export async function fmtOutPutReport(document: vscode.TextDocument): Promise<vo
         fmtStart: 0,
         fmtEnd: document.lineCount - 1,
         from: EFormatChannel.byFormatAllFile,
+        cmdTo1_or_2: getFormatConfig().removeFirstCommaCommand,
     }));
 
     const edit: vscode.WorkspaceEdit = new vscode.WorkspaceEdit();