+++ src/provider/FormattingEditOnType/OnTypeFormattingEditProvider.ts
@@ -1,4 +1,5 @@
 import * as vscode from 'vscode';
+import { getFormatConfig } from '../../configUI';
 import { EFormatChannel } from '../../globalEnum';
 import { FormatCore } from '../Format/FormatProvider';
 import type { TFmtCore, TFmtCoreMap } from '../Format/FormatType';
@@ -18,6 +19,7 @@ export const OnTypeFormattingEditProvider: vscode.OnTypeFormattingEditProvider =
                 fmtStart: position.line - 1,
                 fmtEnd: position.line,
                 from: EFormatChannel.byFormatOnType,
+                cmdTo1_or_2: getFormatConfig().removeFirstCommaCommand,
             });
 
             const e: TFmtCore | undefined = map.get(position.line - 1);