+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
 ## Next v0.0.34(2023-04-XX)
 
+- feat: add format option as `AhkNekoHelp.format.removeFirstCommaCommand`
+- feat: cmd - `format all file` add option to select `AhkNekoHelp.format.removeFirstCommaCommand`
 - feat: add sign of build-in function
 - fix: sign of user-def param is [variadic](https://www.autohotkey.com/docs/v1/Functions.htm#Variadic) case.
 - fix: sign of nested function (4k 90-sec)<https://youtu.be/WcMzNcVWcYA>